Add/Delete Objects Part 1
Added the ability to add/delete objects as well as the ability to minimize the background menu and toggle whether an object can be controlled during play mode.
This task is incomplete as of this push, the next order of business will be to make it so objects cannot spawn on top of or be dragged onto each other while in editor mode.